Reintroduced inverse fields, but this time it works with bulk updates.

+++ presto-couchdb/src/main/java/net/ontopia/presto/spi/impl/couchdb/CouchUpdate.java
@@ -3,6 +3,9 @@ package net.ontopia.presto.spi.impl.couchdb;
 import java.util.Collection;
 import java.util.HashSet;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import net.ontopia.presto.spi.PrestoField;
 import net.ontopia.presto.spi.PrestoSchemaProvider;
 import net.ontopia.presto.spi.PrestoTopic;
@@ -11,11 +14,15 @@ import net.ontopia.presto.spi.PrestoUpdate;
 
 public class CouchUpdate implements PrestoUpdate, CouchChangeSet.CouchChange {
 
+    private static Logger log = LoggerFactory.getLogger(CouchUpdate.class.getName());
+
     private final CouchChangeSet changeSet;
 
     private CouchTopic topic;
     private final PrestoType type;
     private final boolean isNew;
+
+    private int updateCount = 0;
 
     CouchUpdate(CouchChangeSet changeSet, PrestoType type) {
         this(changeSet, null, type);
@@ -38,7 +45,6 @@ public class CouchUpdate implements PrestoUpdate, CouchChangeSet.CouchChange {
         } else {
             this.topic = topic;
         }
-
     }
     
     public Type getType() {
@@ -52,7 +58,11 @@ public class CouchUpdate implements PrestoUpdate, CouchChangeSet.CouchChange {
     public PrestoTopic getTopicAfterUpdate() {
         return topic.getDataProvider().getTopicById(topic.getId());
     }
-    
+
+    public boolean hasUpdate() {
+        return isNew || updateCount > 0;
+    }
+
     public void setValues(PrestoField field, Collection<?> values) {
         Collection<Object> existingValues = topic.getValues(field);
         Collection<Object> remValues = new HashSet<Object>(existingValues);          
@@ -65,10 +75,10 @@ public class CouchUpdate implements PrestoUpdate, CouchChangeSet.CouchChange {
             }
         }
 
-        topic.setValue(field, values);
-        handleFieldAddValues(field, addValues);
-        handleFieldRemoveValues(field, remValues);
-        handleFieldUpdated(field);
+        if (!remValues.isEmpty() || !addValues.isEmpty()) {
+            topic.setValue(field, values);
+            handleFieldUpdated(field, addValues, remValues);
+        }
     }
 
     public void addValues(PrestoField field, Collection<?> values) {
@@ -76,45 +86,49 @@ public class CouchUpdate implements PrestoUpdate, CouchChangeSet.CouchChange {
     }
 
     public void addValues(PrestoField field, Collection<?> values, int index) {
-        topic.addValue(field, values, index);
-        handleFieldAddValues(field, values);
-        handleFieldUpdated(field);
+        if (!values.isEmpty()) {
+            topic.addValue(field, values, index);
+            handleFieldUpdated(field, values, null);
+        }
     }
 
     public void removeValues(PrestoField field, Collection<?> values) {
-        topic.removeValue(field, values);
-        handleFieldRemoveValues(field, values);
-        handleFieldUpdated(field);
+        if (!values.isEmpty()) {
+            topic.removeValue(field, values);
+            handleFieldUpdated(field, null, values);
+        }
     }
 
-    private void handleFieldUpdated(PrestoField field) {
+    private void handleFieldUpdated(PrestoField field, Collection<?> addValues, Collection<?> remValues) {
+        updateCount++;
+
         // update name property
         if (field.isNameField()) {
             topic.updateNameProperty(topic.getValues(field));
         }
-    }
-    
-    private void handleFieldAddValues(PrestoField field, Collection<?> values) {
-//        changeSet.addInverseFieldValue(isNew, topic, field, values);              
-    }
-
-    private void handleFieldRemoveValues(PrestoField field, Collection<?> values) {
-        if (field.isReferenceField() && field.isCascadingDelete()) {
-            // perform cascading delete
-            for (Object value : values) {
-                PrestoTopic rtopic = (PrestoTopic)value;
-                PrestoType rtype = getSchemaProvider().getTypeById(rtopic.getTypeId());
-                if (rtype.isRemovableCascadingDelete()) {
-                    changeSet.deleteTopic(rtopic, rtype);
-                }
-            }
-//        } else {
-//            changeSet.removeInverseFieldValue(isNew, topic, field, entry.getValue());
+        
+        if (addValues != null && !addValues.isEmpty()) {
+            log.info("+" + topic.getId() + " " + field.getId() + " " + addValues);
+            changeSet.addInverseFieldValue(isNew, topic, field, addValues);              
         }
         
+        if (remValues != null && !remValues.isEmpty()) {
+            log.info("-" + topic.getId() + " " + field.getId() + " " + remValues);
+            if (field.isReferenceField() && field.isCascadingDelete()) {
+                // perform cascading delete
+                for (Object value : remValues) {
+                    PrestoTopic rtopic = (PrestoTopic)value;
+                    PrestoType rtype = getSchemaProvider().getTypeById(rtopic.getTypeId());
+                    if (rtype.isRemovableCascadingDelete()) {
+                        changeSet.deleteTopic(rtopic, rtype);
+                    }
+                }
+            } else {
+                changeSet.removeInverseFieldValue(isNew, topic, field, remValues);
+            }
+        }        
     }
-    
-    
+        
     private PrestoSchemaProvider getSchemaProvider() {
         return type.getSchemaProvider();
     }
